+++ src/main.rs
@@ -1,5 +1,6 @@
 use water_http::server::{ ServerConfigurations};
 use water_http::{InitControllersRoot, WaterController};
+
 type MainHolderType = u8;
 InitControllersRoot!{
     name:MAIN_ROOT,
@@ -13,33 +14,45 @@ async fn main() {
     {
         let subscriber  = tracing_subscriber::FmtSubscriber::builder()
             .with_max_level(tracing::Level::DEBUG)
-            .with_env_filter(tracing_subscriber::EnvFilter::new("debug"))
-            .finish()
-            ;
+            .finish();
         tracing::subscriber::set_global_default(subscriber)
             .expect("no thing");
     }
-    let  config = ServerConfigurations::bind("127.0.0.1",8084);
-    water_http::RunServer!(
+   _= tokio::spawn(async move {
+       let  config = ServerConfigurations::bind("127.0.0.1",8084);
+       water_http::RunServer!(
         config,
         MAIN_ROOT,
         MainController
     );
+   }).await;
 }
 WaterController! {
     holder -> crate::MainHolderType,
     name -> MainController,
     functions -> {
         GET => / => main(context) async {
-            _= context.send_status_code_as_final_response(
-                http::status_code::HttpStatusCode::NOT_FOUND
+            _= context.send_str(
+                "hello world"
             ).await;
         },
         POST => / => post(context) async {
-
+            let mut getter = context
+            .getter();
+            let body_chunks_reader =  getter.get_body_by_mechanism(ParsingBodyMechanism::ChunkedTransferEncoding).await;
+            if let ParsingBodyResults::Chunked(IBodyChunks::Chunked(mut reader)) =
+            body_chunks_reader {
+                _= reader.on_chunk_detected(|c,data|{
+                    println!("chunk {} {} {}",c.chunk_size,c.index,String::from_utf8_lossy(data).len());
+                    return Ok(None);
+                }).await;
+            }
             _= context.send_str("hello world").await;
         }
     }
+    extra_code->(..{
+
+    })
 }