+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.0.11
  - new send json method to context to make it much easier to use
  - new send status code method to send http status code as final response to the client inside context
-
+ - multipart form data now support all arbitrary field headers 
 ## 2.0.10 
  - fixing linux kernel system call implementation in one of crates